move static queries to neon-law package
