+++ packages/neon-law/src/content/justice-for-rickie-slaughter.mdx
@@ -6,7 +6,7 @@ updatedAt: "2020-07-01"
 tags: pro-bono,black-lives-matter,criminal-justice
 ---
 
-import { Image } from '@neonlaw/shared-ui/src/components/image';
+import { Image } from '../components/image';
 import { VegasDirections } from '@neonlaw/shared-ui/src/components/vegasDirections';
 
 > Rickie shared an update on July 1, 2020, thanking all of you for checking
