+++ packages/neon-law/src/content/practice-areas.mdx
@@ -6,7 +6,7 @@ tags: neon-law,practice-areas
 description: The practice areas our firm specalizes in.
 ---
 
-import { Image } from '@neonlaw/shared-ui/src/components/image';
+import { Image } from '../components/image';
 
 <Flex flexWrap="wrap">
