+++ packages/neon-law/src/pages/about-us.mdx
@@ -1,7 +1,7 @@
 import { FaLinkedinIn } from 'react-icons/fa';
 import { MdEmail } from 'react-icons/md';
-import { Image } from '@neonlaw/shared-ui/src/components/image';
-import { Seo } from '@neonlaw/shared-ui/src/components/seo';
+import { Image } from '../components/image';
+import { Seo } from '../components/seo';
 
 <PublicLayout>
