+++ packages/neon-law/src/pages/index.mdx
@@ -1,5 +1,5 @@
-import { Image } from '@neonlaw/shared-ui/src/components/image';
-import { Seo } from '@neonlaw/shared-ui/src/components/seo';
+import { Image } from '../components/image';
+import { Seo } from '../components/seo';
 
 <PublicLayout>
