+++ packages/shared/src/components/beginUpwardMobilityQuestionnaire.tsx
@@ -1,7 +1,6 @@
 import { Heading, Text } from '@chakra-ui/core';
 import { Button } from '../components/button';
 import React from 'react';
-import { Seo } from '../components/seo';
 import { decisionTree } from '../components/upwardMobilityQuestions';
 import { navigate } from 'gatsby';
 import { useIntl } from 'gatsby-plugin-intl';
@@ -12,8 +11,6 @@ export const BeginUpwardMobilityQuestionnaire = () => {
   const intl = useIntl();
   return (
     <>
-      <Seo title="Upward Mobility Questionnaire" />
-
       <Heading textAlign="center">
         {intl.formatMessage({ id: 'beginUpwardMQ.heading' })}
       </Heading>
